Relax requirements for deltapsi/heterogen groups, convenience flags
Previously --names and -grp2 flags were required.
These are now optional.
There are default names for quantification groups (grp1 and grp2).
If grp2 is not specified, it is assumed that an intragroup comparison is
desired. Experiments for the two groups will be chosen randomly.

Additionally, new flag --downsample2 enables automatic downsampling of
grp2 to (at most) number of experiments of grp1, as a convenience.